requirements.txt
# Instagram Automation
instagrapi==2.1.2

# Telegram Bot
python-telegram-bot==21.0.1

# Database
mysql-connector-python==8.3.0
PyMySQL==1.1.0

# Task Scheduling
APScheduler==3.10.4

# Security
cryptography==42.0.5
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0
requests==2.31.0

# Logging
coloredlogs==15.0.1
